afi-core/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai     # Gemini
chromadb                # RAG
langchain               # Orquestación RAG
langchain-community
langchain-text-splitters
pypdf                   # Leer PDFs (Libros)
PyPDF2                  # Soporte explícito para ingestión
apscheduler             # Proactividad
imap-tools              # Leer correos
edge-tts                # Voz del Oráculo
matplotlib              # Gráficos
kaleido                 # Exportar gráficos a imagen
pandas
numpy
openpyxl
xlsxwriter
httpx
psycopg2-binary
pytest
tabulate